预测 预测销售特征总数为1校验

+++ src/main/java/com/olivia/peanut/aps/api/entity/apsGoodsForecastUserSaleData/ApsGoodsForecastUserSaleDataDto.java
@@ -4,6 +4,8 @@ import com.olivia.peanut.portal.api.entity.BaseEntityDto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 /**
  * (ApsGoodsForecastUserSaleData)查询对象返回
  *
@@ -19,18 +21,18 @@ public class ApsGoodsForecastUserSaleDataDto extends BaseEntityDto {
   private Long forecastId;
   private Long saleConfigId;
   private Integer year;
-  private Double month_01;
-  private Double month_02;
-  private Double month_03;
-  private Double month_04;
-  private Double month_05;
-  private Double month_06;
-  private Double month_07;
-  private Double month_08;
-  private Double month_09;
-  private Double month_10;
-  private Double month_11;
-  private Double month_12;
+  private BigDecimal month_01;
+  private BigDecimal month_02;
+  private BigDecimal month_03;
+  private BigDecimal month_04;
+  private BigDecimal month_05;
+  private BigDecimal month_06;
+  private BigDecimal month_07;
+  private BigDecimal month_08;
+  private BigDecimal month_09;
+  private BigDecimal month_10;
+  private BigDecimal month_11;
+  private BigDecimal month_12;
 
 }
